pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "MyProject using crewAI"
authors = [{ name = "Muhammad Ubaid", email = "ukubaid@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.100.1,<1.0.0",
    "streamlit>=1.32.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1"
]

[project.scripts]
myproject = "myproject.main:run"
run_crew = "myproject.main:run"
train = "myproject.main:train"
replay = "myproject.main:replay"
test = "myproject.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/myproject"]

[tool.crewai]
type = "crew"
